Extract FileUrlService superclass
